add grpc server

+++ modulesV2/locations/internal/dto.go
@@ -16,7 +16,7 @@ type Location struct {
 type CreatePayload struct {
 	PersonID   int    `json:"personId" db:"person_id"`
 	Coordinate string `json:"coordinate" db:"coordinate"`
-	CreationTS int16  `json:"creationTime"`
+	CreationTS int64  `json:"creationTime"`
 }
 
 type HealthResponse struct {
